requirements.txt
tensorflow==2.16.1
scikit-learn==1.4.2
Pillow==10.3.0
streamlit==1.35.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4